+++ varifier/vcf_evaluate.py
@@ -135,6 +135,7 @@ def evaluate_vcf(
     global_align_min_coord=0,
     global_align_max_coord=float("inf"),
     ignore_non_acgt=True,
+    use_mafft=False,
 ):
     if force:
         subprocess.check_output(f"rm -rf {outdir}", shell=True)
@@ -203,6 +204,7 @@ def evaluate_vcf(
         global_align_min_coord=global_align_min_coord,
         global_align_max_coord=global_align_max_coord,
         ignore_non_acgt=ignore_non_acgt,
+        use_mafft=use_mafft,
     )
     if ref_mask_bed_file is not None:
         logging.info("Masking recall VCF...")